requirements.txt
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.5
django-silk==5.4.1
djangorestframework==3.16.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
gprof2dot==2025.4.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2==2.9.10
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
